separate calculating seed from printing
