+++ src/lib.rs
@@ -6,8 +6,6 @@ use std::thread;
 use std::time::Duration;
 use std::thread::JoinHandle;
 
-const ADDRESS_SIZE: usize = 35;
-
 pub fn main_thing(threads_count: u32, looking_for: &[String]) {
     let mut worlds: HashMap<String, usize> = HashMap::new();
     for looking in looking_for.iter() {
@@ -35,22 +33,10 @@ fn run_some_work(worlds: &HashMap<String, usize>) -> JoinHandle<()> {
         let thread_id = thread::current().id();
         println!("{:?} thread starts with input: {:?}", thread_id, map_copy);
         loop {
-            some_work(&map_copy);
+            let seed_with_address = seed_generator::generate_seed_with_address_for_worlds(&map_copy);
+            println!("seed: {}\naddress: {}\n----------", seed_with_address.seed, seed_with_address.address);
         }
     })
 }
 
-fn some_work(worlds: &HashMap<String, usize>) -> () {
-    let seed_with_address = seed_generator::generate_seed_with_address();
-    if check_address(worlds, &seed_with_address.address) {
-        println!("seed: {}\naddress: {}\n----------", seed_with_address.seed, seed_with_address.address)
-    }
-}
 
-fn check_address(worlds: &HashMap<String, usize>, address: &str) -> bool {
-    worlds.iter().any(|kv| check_concrete_address(address, kv.0, kv.1))
-}
-
-fn check_concrete_address(address: &str, case: &str, case_len: &usize) -> bool {
-    &address[(ADDRESS_SIZE - case_len)..].to_lowercase() == case
-}
